backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
httpx==0.25.2
openai==1.3.7
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
aiofiles==23.2.1
loguru==0.7.2
